Rename FastFileToolGUI to FastFileExtractor
Rename the FastFileToolGUI project and artifacts to FastFileExtractor across the solution and CI. Updated project file (AssemblyName/RootNamespace/Version kept), source namespaces, UI text, and startup log message. Updated .sln to add the new project entries and removed the old project, adjusted .gitignore, README, and GitHub Actions workflows to build/upload publish-extractor artifacts instead of publish-tool. This aligns names and CI artifacts to the new "Extractor" branding.
