+++ FastFileExtractor/FastFileExtractor.csproj
@@ -7,6 +7,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <SelfContained>false</SelfContained>
+    <AssemblyName>FastFileExtractor</AssemblyName>
+    <RootNamespace>FastFileExtractor</RootNamespace>
     <Version>3.1.0</Version>
     <AssemblyVersion>3.1.0</AssemblyVersion>
     <FileVersion>3.1.0</FileVersion>
